AlbumAdapter, AlbumRecyclerViewInterface, ArtistAdapter, ArtistRecyclerViewInterface, Song_RecyclerViewAdapter, and SongRecyclerViewInterface: moved to viewHolders package to keep things organized, checked first to make sure there are no access modifier issues.
