+++ app/src/main/java/com/example/spawtify/viewHolders/AlbumRecyclerViewInterface.java
@@ -1,4 +1,4 @@
-package com.example.spawtify;
+package com.example.spawtify.viewHolders;
 
 public interface AlbumRecyclerViewInterface {
     void onItemClick(int position);
