+++ app/src/main/java/com/example/spawtify/viewHolders/ArtistRecyclerViewInterface.java
@@ -1,4 +1,4 @@
-package com.example.spawtify;
+package com.example.spawtify.viewHolders;
 
 public interface ArtistRecyclerViewInterface {
     void onItemClick(int position);
