+++ app/src/main/java/com/example/spawtify/viewHolders/SongRecyclerViewInterface.java
@@ -1,4 +1,4 @@
-package com.example.spawtify;
+package com.example.spawtify.viewHolders;
 
 public interface SongRecyclerViewInterface {
     void onItemClick(int position);
