+++ app/src/main/java/com/example/spawtify/viewHolders/Song_RecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.spawtify;
+package com.example.spawtify.viewHolders;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,6 +8,9 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.spawtify.R;
+import com.example.spawtify.SongModel;
 
 import java.util.List;
